+++ test/mcrypto/test_util.py
@@ -9,7 +9,6 @@ from forge_sdk import protos
 class DIDUtilTest(unittest.TestCase):
 
     def test_get_asset_address(self):
-        account = "z1cYzB4LHUKs7i6ZYt7BHRVm4eofnHeutoL"
-        itx = protos.CreateAssetTx(data=Any(value=b'123'))
-        asset_address = did.get_asset_address(account, itx)
-        assert (asset_address == "zjdjoBZ8TCyQhbY2q3GZfJFAcWWBbF8ZXrts")
+        itx = protos.CreateAssetTx()
+        asset_address = did.get_asset_address(itx)
+        assert asset_address == "zjdtErp27MYs7BuaoQWBF6DSg6jU4DrE1VQ2"